+++ components/ModMainContent.vue
@@ -1,5 +1,12 @@
 <template>
     <div class="h-full overflow-hidden">
+        <Modal
+            data-testid="moderation-modal"
+            class=" min-h-20 min-w-20 fixed top-0 left-0 flex items-center justify-center h-full w-full z-10
+        bg-secondary bg-opacity-40"
+        >
+            <ModerationConfirmation />
+        </Modal>
         <div v-if="screenStore.activeScreen === ModerationScreen.Dashboard">
             <ModSubmissionListContainer />
         </div>
