+++ i18n/locales/en.json
@@ -166,7 +166,9 @@
         "addHealthCareProfessionalLocaleName": "Add Name by Locale",
         "addresses": "Addresses",
         "clickToRemoveHealthCareProfessionalName": "Click on a row to remove a name",
-        "confirmationButton": "Leave Page",
+        "confirmLeavePageButton": "Yes, leave page",
+        "confirmApproveSubmissionMessage": "Are you sure you want to approve this submission?",
+        "confirmApproveSubmissionButton": "Yes, approve submission",
         "hasUnsavedChanges": "You have unsaved changes. Are you sure you want to leave?",
         "contactInformation": "Contact Information",
         "facilityHeading": "Facility Information",
