+++ i18n/locales/pt.json
@@ -213,7 +213,7 @@
     "selectSpecialties": "Specialties",
     "selectDegrees": "Degrees",
     "selectLocales": "Select Locales Spoken",
-    "confirmationButton": "Leave Page",
+    "confirmLeavePageButton": "Yes, leave page",
     "hasUnsavedChanges": "You have unsaved changes. Are you sure you want to leave?"
   },
   "login": {
